# static/api.yaml
openapi: '3.0.3'
info:
  version: '1.1.0'
  title: Netsoc webspaced
  description: >
    API for managing next-gen webspaces.
servers:
  - url: https://webspaced.netsoc.ie/v1
  - url: https://webspaced.staging.netsoc.ie/v1
  - url: http://localhost:8081/v1
components:
  securitySchemes:
    jwt:
      $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/securitySchemes/jwt'
    jwt_admin:
      $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/securitySchemes/jwt_admin'
  parameters:
    Domain:
      name: domain
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Domain'
    ExternalPort:
      name: ePort
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Port'
    InternalPort:
      name: iPort
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Port'
  responses:
    InternalError:
      description: General server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Resource does not exist (e.g. user, webspace)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error (e.g. Required field missing)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
    ConflictError:
      description: Webspace for username already exists / is already running
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Message explaining the error
...
    Domain:
      type: string
      description: Custom domain
      example: example.com
    Domains:
      type: array
      items:
        $ref: '#/components/schemas/Domain'
      description: List of webspace custom domains
    Port:
      type: integer
      format: int32
      description: Network port
      example: 8080
...
    Usage:
      type: object
      required:
        - cpu
        - disks
        - memory
        - processes
      description: Website resource usage
      properties:
        cpu:
          type: integer
          format: int64
          description: CPU time (nanoseconds)
          example: 685502875
        disks:
          type: object
          additionalProperties:
            type: integer
            format: int64
            description: Disk usage in bytes
          example:
            root: 16777216
        memory:
          type: integer
          format: int64
          description: Memory usage in bytes
          example: 10452992
        processes:
          type: integer
          format: int64
          description: Number of processes
          example: 8
    InterfaceCounters:
      type: object
      required:
        - bytesReceived
        - bytesSent
        - packetsReceived
        - packetsSent
      description: Counters for a network interface
      properties:
        bytesReceived:
          type: integer
          format: int64
          example: 46897
        bytesSent:
          type: integer
          format: int64
          example: 9910
        packetsReceived:
          type: integer
          format: int64
          example: 122
        packetsSent:
          type: integer
          format: int64
          example: 46
    InterfaceAddress:
      type: object
      required:
        - family
        - address
        - netmask
      description: Network interface address
      properties:
        family:
          type: string
          default: inet
        address:
          type: string
          example: '10.233.0.143'
        netmask:
          type: string
          example: '24'
        scope:
          type: string
          example: global
    NetworkInterface:
      type: object
      required:
        - mac
        - mtu
        - state
        - counters
        - addresses
      description: Webspace network interface status
      properties:
        mac:
          type: string
          description: MAC address
          example: '00:16:3e:34:9a:d4'
        mtu:
          type: integer
          default: 1500
        state:
          type: string
          default: up
        counters:
          $ref: '#/components/schemas/InterfaceCounters'
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/InterfaceAddress'
    State:
      type: object
      required:
        - running
        - uptime
        - usage
        - networkInterfaces
      description: Webspace state
      properties:
        running:
          type: boolean
          default: false
          example: true
        uptime:
          type: number
          format: double
          description: Length of time for which container has been running (seconds)
        usage:
          $ref: '#/components/schemas/Usage'
        networkInterfaces:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/NetworkInterface'
          example:
            eth0:
              mac: '00:16:3e:34:9a:d4'
              mtu: 1500
              state: up
              counters:
                bytesReceived: 46897
                bytesSent: 9910
                packetsReceived: 122
                packetsSent: 46
              addresses:
                - family: inet
                  address: 10.233.0.143
                  netmask: '24'
                  scope: global
                - family: inet6
                  address: 'fd42:7373:3d88:e37a:216:3eff:fe34:9ad4'
                  netmask: '64'
                  scope: global
                - family: inet6
                  address: 'fe80::216:3eff:fe34:9ad4'
                  netmask: '64'
                  scope: link
...
paths:
...
  /webspace/{username}/state:
    get:
      summary: Retrieve webspace state
      operationId: getState
      tags: [state]
      parameters:
        - $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/parameters/UsernameOrSelf'
      security:
        - jwt: []
        - jwt_admin: []
      description: >
        Retrieve webspace state
      responses:
        '200':
          description: Webspace state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/State'
              examples:
                running:
                  description: Running
                  value:
                    running: true
                    uptime: 32
                    usage:
                      cpu: 685502875
                      disks: {}
                      memory: 10452992
                      processes: 8
                    networkInterfaces:
                      eth0:
                        mac: '00:16:3e:34:9a:d4'
                        mtu: 1500
                        state: up
                        counters:
                          bytesReceived: 46897
                          bytesSent: 9910
                          packetsReceived: 122
                          packetsSent: 46
                        addresses:
                          - family: inet
                            address: '10.233.0.143'
                            netmask: '24'
                            scope: global
                          - family: inet6
                            address: 'fd42:7373:3d88:e37a:216:3eff:fe34:9ad4'
                            netmask: '64'
                            scope: global
                          - family: inet6
                            address: 'fe80::216:3eff:fe34:9ad4'
                            netmask: '64'
                            scope: link
                notRunning:
                  description: Not running
                  value:
                    running: false
                    uptime: 0
                    usage:
                      cpu: 0
                      disks: {}
                      memory: 0
                      processes: 0
                    networkInterfaces: {}
        '401':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AuthError'
        '403':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AdminError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      summary: Re-generate webspace backend config
      operationId: sync
      tags: [state]
      parameters:
        - $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/parameters/UsernameOrSelf'
      security:
        - jwt: []
        - jwt_admin: []
      description: Forces a reload of reverse proxy and port forwarding configuration
      responses:
        '204':
          description: No content
        '401':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AuthError'
        '403':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AdminError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Start webspace container
      operationId: start
      tags: [state]
      parameters:
        - $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/parameters/UsernameOrSelf'
      security:
        - jwt: []
        - jwt_admin: []
      responses:
        '204':
          description: No content
        '401':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AuthError'
        '403':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AdminError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      summary: Reboot webspace container
      operationId: reboot
      tags: [state]
      parameters:
        - $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/parameters/UsernameOrSelf'
      security:
        - jwt: []
        - jwt_admin: []
      responses:
        '204':
          description: No content
        '401':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AuthError'
        '403':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AdminError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Shut down webspace container
      operationId: shutdown
      tags: [state]
      parameters:
        - $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/parameters/UsernameOrSelf'
      security:
        - jwt: []
        - jwt_admin: []
      responses:
        '204':
          description: No content
        '401':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AuthError'
        '403':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AdminError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /webspace/{username}/domains:
    get:
      summary: Retrieve webspace domains
      operationId: getDomains
      tags: [domains]
      parameters:
        - $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/parameters/UsernameOrSelf'
      security:
        - jwt: []
        - jwt_admin: []
      responses:
        '200':
          description: Webspace domains
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domains'
        '401':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AuthError'
        '403':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AdminError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /webspace/{username}/domains/{domain}:
    post:
      summary: Add custom domain
      operationId: addDomain
      tags: [domains]
      parameters:
        - $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/parameters/UsernameOrSelf'
        - $ref: '#/components/parameters/Domain'
      security:
        - jwt: []
        - jwt_admin: []
      description: >
        Domain will be verified by looking for a `TXT` record of the format `webspace:id:<user id>`
      responses:
        '201':
          description: No content
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AuthError'
        '403':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AdminError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Delete custom domain
      operationId: removeDomain
      tags: [domains]
      parameters:
        - $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/parameters/UsernameOrSelf'
        - $ref: '#/components/parameters/Domain'
      security:
        - jwt: []
        - jwt_admin: []
      responses:
        '204':
          description: No content
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AuthError'
        '403':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AdminError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /webspace/{username}/ports/{ePort}/{iPort}:
    post:
      summary: Add port forward
      operationId: addPort
      tags: [ports]
      parameters:
        - $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/parameters/UsernameOrSelf'
        - $ref: '#/components/parameters/ExternalPort'
        - $ref: '#/components/parameters/InternalPort'
      security:
        - jwt: []
        - jwt_admin: []
      responses:
        '201':
          description: No content
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AuthError'
        '403':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AdminError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /webspace/{username}/ports/{ePort}:
    delete:
      summary: Delete port forward
      operationId: removePort
      tags: [ports]
      parameters:
        - $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/parameters/UsernameOrSelf'
        - $ref: '#/components/parameters/ExternalPort'
      security:
        - jwt: []
        - jwt_admin: []
      responses:
        '204':
          description: No content
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AuthError'
        '403':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AdminError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /webspace/{username}/log:
    get:
      summary: Retrieve webspace console log
      operationId: getLog
      tags: [console]
      parameters:
        - $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/parameters/UsernameOrSelf'
      security:
        - jwt: []
        - jwt_admin: []
      responses:
        '200':
          description: Console log
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AuthError'
        '403':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AdminError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Clear webspace console log
      operationId: clearLog
      tags: [console]
      parameters:
        - $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/parameters/UsernameOrSelf'
      security:
        - jwt: []
        - jwt_admin: []
      responses:
        '204':
          description: No content
        '401':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AuthError'
        '403':
          $ref: 'https://raw.githubusercontent.com/netsoc/iam/master/static/api.yaml#/components/responses/AdminError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
...